+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 - Lobby / Waiting room move button / password (!852)
 - Add unread indicator for chat messages (!844)
+- Add legal vote token copy field. (!836)
 
 ## 1.2.0